LayoutTests/imported/w3c/web-platform-tests/webrtc/protocol/h264-profile-levels.https_interop-2026-expected.txt


FAIL Level 4 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 8192 but got 14400
PASS Level 5 H264 video is appropriately constrained
FAIL Level 3.1 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 3600 but got 14400
FAIL Level 3.2 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 5120 but got 14400
FAIL Level 4.1 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 8192 but got 14400
FAIL Level 4.2 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 8704 but got 14400
PASS Level 5.1 H264 video is appropriately constrained
PASS Level 5.2 H264 video is appropriately constrained
